requirements.txt
amqp==5.2.0
asgiref==3.4.1
async-timeout==4.0.3; python_full_version < '3.11.3'
attrs==21.2.0
autopep8==1.6.0
backports.zoneinfo==0.2.1; python_version < '3.9'
billiard==4.2.0
celery==5.4.0
certifi==2021.10.8
charset-normalizer==2.0.10; python_version >= '3'
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
click==8.1.7
cloudinary==1.28.0
coverage==6.2
dj-database-url==0.5.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.10.0
django-filter==21.1
django-heroku==0.3.1
django-rest-passwordreset==1.2.1
django==3.2.9
djangorestframework-simplejwt==5.0.0
djangorestframework==3.12.4
drf-spectacular==0.21.1
factory-boy==3.2.1
faker==10.0.0
gunicorn==20.1.0
idna==3.3; python_version >= '3'
importlib-resources==5.4.0; python_version < '3.9'
inflection==0.5.1
jsonschema==4.3.1
kombu==5.3.7
pillow==8.4.0
prompt-toolkit==3.0.47
psycopg2==2.9.2
pycodestyle==2.8.0
pyjwt==2.3.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
pyyaml==6.0
redis==5.0.7
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
typing-extensions==4.12.2; python_version < '3.10'
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.8
vine==5.1.0
wcwidth==0.2.13
whitenoise==5.3.0
zipp==3.6.0; python_version < '3.10'
